cfunction.R: print tail(errmsg) if it is too long
The errfile/msg contains compiler warnings, which will often be many, long, and non-useful if the code uses Boost, Eigen, etc. But R truncates the message at 1000 characters by default starting from the begining, which shows the first few compiler warnings and not the actual error message at the end. So, it is better to show the last few lines than the first few lines.