use `system2` instead of system (#12)
* DESCRIPTION: minor bump of version number

* cfunction.R: use system2 instead of system

* cfunction.R: only print errmsg if there is an error

* cfunction.R: do not print stdout

if someone wants to see the compiler call, they can do that by specifying verbose = TRUE

* cfunction.R: print tail(errmsg) if it is too long

The errfile/msg contains compiler warnings, which will often be many, long, and non-useful if the code uses Boost, Eigen, etc. But R truncates the message at 1000 characters by default starting from the begining, which shows the first few compiler warnings and not the actual error message at the end. So, it is better to show the last few lines than the first few lines.

* cfunction.R: use sum(nchar(errmsg))

since errmsg is a character vector of multiple lines

* try harder to get the make cmds

* cfunction.R: do not setwd(wd)

because there is already a on.exit(setwd(wd))

* cfunction.R: restore accidentally deleted {

* cfunction.R: adjust error and verbose reporting

facilitate showing what make would do via --dry-run

do not setwd() because it is already in an on.exit()

account for the limits on printing the error message, which may be
thousands of lines long due to compiler warnings from Eigen / Boost

* importFrom("utils", "tail")

* Revert "DESCRIPTION: minor bump of version number"

This reverts commit 889c96578be38c43495b259e7d112ae518fb8ae8.