Correct path generation
This reverts the generation of the 'new_path' of the DLL to the version
where `normalizePath` calls `file.path`. These paths can then correctly
be matched to the normalized paths of the active DLLs also on Windows,
so that unloading of active DLL works as intended. The argument
'mustWork = FALSE' to 'normalizePath' gets rid of the warnings which we
do not care for as we know that the new path will often not point to an
existing file at this point.